scripts: Add glue script for multiarch images
The glue script is used to create a multiarch image from
multiple images built for different architectures. It pulls the images
from the registry, creates a manifest list, and pushes the resulting
multiarch image back to the registry.

This is useful for projects that was built in different architectures
and machines, and need to be available for different platforms.

Signed-off-by: Matheus Castello <matheus@castello.eng.br>